Framework: split the sites selector into multiple files
