+++ client/state/sites/selectors/index.js
@@ -1,0 +1,15 @@
+/** @ssr-ready **/
+
+// These could be rewritten as `export * from`, pending resolution of Babel
+// transform bug: http://phabricator.babeljs.io/T2877
+import * as jetpack from './jetpack';
+import * as plan from './plan';
+import * as site from './site';
+import * as sites from './sites';
+
+export default {
+	...jetpack,
+	...plan,
+	...site,
+	...sites
+};
