+++ client/state/sites/selectors/jetpack.js
@@ -1,0 +1,72 @@
+/** @ssr-ready **/
+
+/**
+ * External dependencies
+ */
+import { includes } from 'lodash';
+
+/**
+ * Internal dependencies
+ */
+import { getRawSite, getSiteOption } from './site';
+import versionCompare from 'lib/version-compare';
+
+/**
+ * Returns true if site is a Jetpack site, false if the site is hosted on
+ * WordPress.com, or null if the site is unknown.
+ *
+ * @param  {Object}   state  Global state tree
+ * @param  {Number}   siteId Site ID
+ * @return {?Boolean}        Whether site is a Jetpack site
+ */
+export function isJetpackSite( state, siteId ) {
+	const site = getRawSite( state, siteId );
+	if ( ! site ) {
+		return null;
+	}
+
+	return site.jetpack;
+}
+
+/**
+ * Returns true if the site is a Jetpack site with the specified module active,
+ * or false if the module is not active. Returns null if the site is not known
+ * or is not a Jetpack site.
+ *
+ * @param  {Object}   state  Global state tree
+ * @param  {Number}   siteId Site ID
+ * @param  {String}   slug   Module slug
+ * @return {?Boolean}        Whether site has Jetpack module active
+ */
+export function isJetpackModuleActive( state, siteId, slug ) {
+	const modules = getSiteOption( state, siteId, 'active_modules' );
+	if ( ! modules ) {
+		return null;
+	}
+
+	return includes( modules, slug );
+}
+
+/**
+ * Returns true if the Jetpack site is running a version meeting the specified
+ * minimum, or false if the Jetpack site is running an older version. Returns
+ * null if the version cannot be determined or if not a Jetpack site.
+ *
+ * @param  {Object}   state   Global state tree
+ * @param  {Number}   siteId  Site ID
+ * @param  {String}   version Minimum version
+ * @return {?Boolean}         Whether running minimum version
+ */
+export function isJetpackMinimumVersion( state, siteId, version ) {
+	const isJetpack = isJetpackSite( state, siteId );
+	if ( ! isJetpack ) {
+		return null;
+	}
+
+	const siteVersion = getSiteOption( state, siteId, 'jetpack_version' );
+	if ( ! siteVersion ) {
+		return null;
+	}
+
+	return versionCompare( siteVersion, version, '>=' );
+}
