+++ client/state/sites/selectors/plan.js
@@ -1,0 +1,91 @@
+/** @ssr-ready **/
+
+/**
+ * External dependencies
+ */
+import { get } from 'lodash';
+
+/**
+ * Internal dependencies
+ */
+import { getRawSite } from './site';
+
+/**
+ * Returns a site's plan object by site ID.
+ *
+ * The difference between this selector and sites/plans/getPlansBySite is that the latter selectors works
+ * with the /sites/$site/plans endpoint while the former selectors works with /sites/$site endpoint.
+ * Query these endpoints to see if you need the first or the second one.
+ *
+ * @param  {Object}  state  Global state tree
+ * @param  {Number}  siteId Site ID
+ * @return {?Object}        Site's plan object
+ */
+export function getSitePlan( state, siteId ) {
+	const site = getRawSite( state, siteId );
+
+	if ( ! site ) {
+		return null;
+	}
+
+	if ( get( site.plan, 'expired', false ) ) {
+		if ( site.jetpack ) {
+			return {
+				product_id: 2002,
+				product_slug: 'jetpack_free',
+				product_name_short: 'Free',
+				free_trial: false,
+				expired: false
+			};
+		}
+
+		return {
+			product_id: 1,
+			product_slug: 'free_plan',
+			product_name_short: 'Free',
+			free_trial: false,
+			expired: false
+		};
+	}
+
+	return site.plan;
+}
+
+/**
+ * Returns true if the current site plan is a paid one
+ *
+ * @param  {Object}   state         Global state tree
+ * @param  {Number}   siteId        Site ID
+ * @return {?Boolean}               Whether the current plan is paid
+ */
+export function isCurrentPlanPaid( state, siteId ) {
+	const sitePlan = getSitePlan( state, siteId );
+
+	if ( ! sitePlan ) {
+		return null;
+	}
+
+	return sitePlan.product_id !== 1 && sitePlan.product_id !== 2002;
+}
+
+/**
+ * Returns true if site is currently subscribed to supplied plan and false otherwise.
+ *
+ * @param  {Object}   state         Global state tree
+ * @param  {Number}   siteId        Site ID
+ * @param  {Number}   planProductId Plan product_id
+ * @return {?Boolean}               Whether site's plan matches supplied plan
+ */
+export function isCurrentSitePlan( state, siteId, planProductId ) {
+	if ( planProductId === undefined ) {
+		return null;
+	}
+
+	const sitePlan = getSitePlan( state, siteId );
+
+	if ( ! sitePlan ) {
+		return null;
+	}
+
+	return sitePlan.product_id === planProductId;
+}
